Support IE 9 and lower better
Although it is already supported, the problem is that the back-end
expects 2 extra metaframes, if a stack trace was generated by us.

The first approach in `generateStacktrace()` provides these frames,
but if it fails, we fall back to the second approach, which lacks
them. It means that on the back-end side the parser chops off two
important frames.