Fix bugsnag not notifying when notifyReleaseStages is set but releaseStage isn't.
As is written here https://bugsnag.com/docs/notifiers/js under releaseStage, by default, the releaseStage is "production". It is true that when notifying, if releaseStage is not specified, the server will consider it as "production".

The problem I noticed is, when notifyReleaseStages is set but releaseStage isn't, then it won't notify at all. That is contrary to what you'd expect when reading the documentation.