Unflake metadata tests
Win7/Firefox 31,32 were occassionally flaking out on the metadata
tests. This is because they expected `request.metadata` to match
the fixture provided, but intermittently the actual result included a
`metadata.script` property which did not exist on the fixture.

Since the tests were already removing a top level property from metaData
(`device`) I updated all of the metadata tests just to do an equality
check on the properties they expected to exist, rather than delete
all of the properties it wasn't expecting.

May the Sauce be with this commit.