Better non-error extraction in window.onerror handler
A non-error thrown and seen in the window.onerror handler didn't get
put into the error message (like it does for notify()) arguments.

This uses the String(arg) function to use whatever was "thrown"
as the error message, then JSON serializes the full value in the
window onerror metadata tab.