Ensure inline script content is added after location change
The plugin that adds inline script content compared the location
of stackframe[0] with the current window.location.href to decide
whether it was an inline script or not.

This check erroneously failed if the window.location.href changed
after page load - via the history.push|replaceState() API. Now the
window's original location is saved and compared against.