Don't check for duplicate breadcrumbs
It was expensive to serialise every item before it was added to
the breadcrumb array, and compare it with a fresh serialisation
of the last breadcrumb added. Turns out the duplicate filtering
would rarely (if ever) have been invoked because the crumbs would
have to have occurred in the same millisecond anyway. In the event
of an infinite (or very repetitive) loop adding crumbs, the max
length configuration ensures we don't use too much memory.

This change results in a ~30x perf improvement of `leaveBreadcrumb()`.