fix(deps): Fork cuid dependency to fix rollup bundling issue
The version of cuid used did not play well with rollup's bundling, due to this issue:
https://github.com/ericelliott/cuid/issues/41. When cuid was added to bugsnag, an update wasn't
available, but v2.x was published in Jan 2018 which fixes this issue. Unfortunately the updates
broke for ie8 in the process (which we need to support) so we are maintaining a fork, which is
published at @bugsnag/cuid.