fix(deps): Switch (u)uid implementation
cuid() was great but it caused problems when bundling with rollup. Newer versions of cuid that have
been published fix the rollup issues but it no longer supports IE8 (which we need). Our options were
to maintain a fork of cuid with the minor change needed to support IE8 or find a different uid
implementation. Since every alternative implementation I tried added bytes to our minified bundle, I
focused on finding the smallest possible uuid id implementation. It wasn't on npm so I added it
directly to the codebase.