feat(transports): Default to https:// for error/session reporting
The default configuration was a protocol relative url "//…" which would match the protocol of the
page. This was to support IE8/9, as they would only send XDomain POST requests to a URL on the same
protocol as the current page. However the relative protocol breaks in a couple of places, namely
developing locally on a file:// url and in Electron development.

This change means that http will
only ever be used in IE8/9 when the protocol of the page is http, OR if the library is manually
configured to point to a non-ssl endpoint.