build(browserify): Add envify transform to compile out test-only code
A destroy() method has been used on plugins that have global side-effects, so that in the tests they
can have their side-effects removed after they have been tested. This pattern is only used in tests
so this code was unnecessary in the production bundle. Using envify we can inject the environment
into the production bundle which then causes uglify to remove the unreachable code between env
checks.