feat(sessions): Enable session tracking by default
This commit enables session tracking by default, and makes some requisite changes to ensure the
relevant endpoints are properly configured. The options `endpoint` and `sessionEndpoint` are
soft-deprecated, and they are replaced by the `endpoints: { notify, sessions }` object.
Previously
the validation framework was context-free – each value was validated indpendently. Now, some options
need to know the state of other options when validating and setting defaults, so this has been
updated accordingly.