fix(config): Ensure validation messages are properly surfaced
When bad configuration caused the validation step to fail, the various error messages were attached
to the error which was then thown. The errors where subsequently passed the client's logger, which
would not output anywhere due to the configuration having failed. This commit contructs an error
that has the messages in its `message` property and doesn't try to catch it.