fix: Don't attempt to parse a stacktrace if it doesn't contain any frames
SyntaxErrors in Chrome have an error.stack property which is just `${err.name}: ${err.message}` and
no stack frames but because the err.stack value is truthy we attempted to parse it with
error-stack-parser. The result of this was that the filename proeprty of the stackframes was not
correct and caused issues with grouping. This change causes Report.getStacktrace(err) to generate a
backtrace instead and then the file/line/col properties get filled in with the values from
window.onerror.