feat: Make exception type dynamic
For the pipeline to differentiate between Node/browser errors
coming from same notifier, the exception type needs to discern
the platform of the error. This is implemented by checking for
the existence of the IS_BROWSER env var. In the browser build,
it is set to a truthy value and replaced using the envify
transform, which is then optimised away in the uglified bundle.