refactor: Remove beforeSession callbacks
This was at best a half baked implementation and the single piece of functionality relying on it –
@bugsnag/plugin-browser-device – was able to use a different mechanism to support augmentation of
the session payload.
The existing implementation caused problems for the session tracking mechanism
in node.
If we want to implement an equivalent of beforeSend for sessions, we should think about
the ramifications and design it properly.