feat(plugin-node-device): Add process.versions to device data
The version of Node was not present anywhere in the payload – what's more there are a whole bunch of
other versions available in the case that Node happened to be compiled with any different binary
internals. The contents of this hash is now added to the device data tab verbatim.