fix(plugin-window-unhandled-rejection): Tolerate property access errors
In the callback for unhandled rejection events, `event.detail` can be any unknown/untrusted object.
In some cases accessing the `reason` property on it has caused a security error – perhaps because
the event detail came from an iframe or something similar. This change wraps that access in a try
catch and adds a unit test for such a scenario.

Fixes #394.