fix(plugin-window-onerror): Support additional onerror call signature
For some events triggered by jQuery, the second argument to onerror can be an object. This
previously caused the onerror handler to misidentify the kind of error that had happened. This fix
amends the check on the "url" argument, only determining it to be a url if it is defined and is a
string.

Fixes #392.