feat(delivery-node): Remove dependency on request
Request is a huge dependency, both in terms of size and number of transient dependencies, compared
with the rest of the node notifier. This commit replaces the sending of error reports via request
with the built-in http/https modules, and exposes the `agent` option to faciliate sending via proxy
(which is the only feature of request that we relied on).