fix(core): Ensure notify() callback is always called
In the following circumstances, the notify callback wasn't called:

- report not sent due to
releaseStage/notifyReleaseStages config
- report not sent due to beforeSend callback

This meant
that the configured onUncaughtException and onUnhandledRejection callbacks (in Node) were never
called, and so unhandled errors were effectively silently swallowed. Although intentionally not
reported, it's still useful to know about so this change ensures the callbacks run. The types were
updated to remove the no-longer present notion that the the return value of notify() indicated
whether the report was sent or not.

Fixes #449