feat(plugin-inline-script-content): Improve trace handler
When the traced callback doesn't error, the original script value can be reset synchronously. In
some cases the asynchronous null-ing of the current script allowed the wrong script to be identified
and attached to reports when an error happened in the same tick as a wrapped callback. Also, while
none of the testing so far yeilded a concern about the wrapped callback returning the original
callback's return value, but it should operate as transparently as possible. This ensures the return
value is passed on.