fix(browser): Stub a .ts file to allow Angular AOT mode to recompile
In Angular's AOT mode, the first full build would work but an incremental build upon making a change
would fail with an error that it couldn't locale the bugsnag browser types module. By generating an
empty bugsnag.ts file in the browser/dist/types directory, Angular is convinced that it's ok and it
all compiles and runs correctly.

Fixes #523