refactor(config): Rename autoCaptureSessions -> autoTrackSessions and simplify validation logic
Per the updated spec, the option to disable sessions is called `autoTrackSessions`. In addition to
making that update, a suggestion made here¹ was incorporated to make the session endpoint always
required (this only applies when endpoints is set). Previously the validation of the
`autoCaptureSessions` field was dependent on whether a session endpoint was provided, which made the
logic unnecessarily complex.

¹https://github.com/bugsnag/bugsnag-js/pull/630/files#r343663933