build: add cucumber test scaffold
Adds a general structure for a set of cucumber features, and example
scenarios, where an app is launched and activates various actions,
eventually sending requests to a mock upload server.

Of particular interest is that unhandled exceptions in the main process
do not crash on windows (??), so we will need some creative expectations
around some of these cases.