fix(plugin-inline-script-content): Check document.readyState in case we missed the interactive event
If Bugsnag is loaded after the readyState=interactive event has fired, it incorrectly assumed that
inline scripts were still executing. This ensures that if the ready event is missed, it doesn't
erroneously grab the last <script> on the page as the currentScript for errors that didn't originate
in an inline script.