Automatically notify when the lambda may timeout
We can never be 100% sure that a timeout will happen because we need
to notify with enough time for the event to be delivered. Therefore
this is a best effort with a generous default value (1000ms) that
errs on the side of delivering the warning at the cost of possible
false positives

The default value can be changed by providing "lambdaTimeoutNotifyMs"
to createHandler. This is useful in cases where the lambda function
is very quick to execute and so can have a very short timeout without
necessarily being at risk of timing out

If set to "0" we don't bother adding the warning at all. This is
largely arbitrary as a value of 0 would never be triggered (because
the lambda would timeout at the same time)