Handle inForeground when all windows are closed
No blur event is triggered for the last window to close. On Linux/
Windows this is fine as the app will quit then anyway, but on macOS
it won't (usually). Therefore we need to track 'window-all-closed'
to handle the case where all windows are closed but the app doesn't
quit