build: make local integration testing handle crash dialog gracefully
When a previous scenario crashes on macOS, the next run of the test app
will launch a dialog asking whether to open previous windows. This
change prevents not responding to the dialog from failing the test (as
a non-response will instead retry the app launch, and the dialog is only
triggered once).