Avoid Breadcrumb.toJSON by copying to an object
The Breadcrumb.toJSON method renames the internal properties of a
breadcrumb object. This is fine, but with some changes the toJSON
method stops being called during serialisation. When this happens,
the breadcrumbs are silently dropped because the 'name' is undefined
and the core Client drops breadcrumbs without a name/message

By avoiding toJSON in all cases, this should be a bit more robust