Verify onBreadcrumb syncing can be cancelled
Our onBreadcrumb callbacks are added to the front of the list of
callbacks. This would usually mean they were also called first, which
would be problematic as then cancelled breadcrumbs would still be
synced. However, we run breadcrumb callbacks in reverse order (last
to first) so they are actually called _last_. This allows breadcrumbs
to be cancelled normally while also cancelling the syncing