feat(react-native): Always upload JS version of bundle file
When Hermes is used, the final main.jsbundle is a precompiled binary
format. This means that the Bugsnag backend can't parse it as JS
to find the original (precompiled) method name.

The JS bundle (which is fed into the Hermes compiler) alway exists in
this location, whether Hermes is in use or not.