Plugin for Azure Functions
This is an initial commit of a @bugsnag/js plugin for Azure Functions. It wraps the Azure Functions handler and catches unhandled exceptions. It also automatically adds the Azure Functions context to any errors reported to Bugsnag.

This plugin is based on the AWS Lambda plugin.

Basic unit tests are included.