Fix issue with context on overlapping invocations
If two function invocations overlap, it is possible the context sent on an error contains incorrect data. Imagine this sequence:

    1. Func 1 starts
    2. Func 2 starts
    3. Func 1 throws an error

In this scenario, since we're using a per-process client, the context reported with the error is a combination of function 1 and function 2 context.

To address this, we treat the scenario similar to the express plugin, and clone the client for each invocation so the metadata associated with it is correct.

This also adds a bugsnag property to the context, which the function can use to notify of errors (and have context included automatically).