Fix a possible crash in app breadcrumbs plugin
This can happen when closing a window immediately after moving it
(i.e. within the ~250ms debounce timer). We protect against this in
two ways:

1. check if the window is destroyed before using it in the 'moved'
   callback
2. cancel a queued call to the 'moved' callback when the 'closed'
   event fires, as this is the point where the browser window is
   destroyed

The other event listeners should never be called with a destroyed
browser window (other than 'closed') because they are not debounced
and so run as soon as the event is emitted